requirements.txt
certifi==2019.6.16
chardet==3.0.4
codecov==2.0.15
configobj==5.0.6
coverage==4.5.4
idna==2.8
ply==3.11
pyhcl==0.3.12
python-dotenv==0.10.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
toml==0.10.0
urllib3==1.25.3
xmltodict==0.12.0
